Add build operation types to log events
Introduces a BuildOperationType enum and adds it to the
default BuildOperationDescriptor. Passes this information
to the logging infrastructure through the BuildOperationExecutor.

This will allow OutputEventListeners to react to output events
from different types of operations differently.

RenderableOutputEvents now have an optional, compact, build
operation descriptor that is a small, serializable subset of a
BuildOperationDescriptor (just an id and operation type).

Issue: #1818